# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GadgetByteNepal",
        "description": "GadgetByte Nepal Rest Apis Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/post-review": {
            "get": {
                "tags": [
                    "Post Review"
                ],
                "summary": "Get a list of Post Reviews",
                "operationId": "1da2c5f13265322183c2aa226cca31a4",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "query",
                        "description": "Post id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search term for filtering by name ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (optional, default: 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostReviewResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Post Review"
                ],
                "summary": "Create a new Post Review",
                "operationId": "684f9fe8aa52aec6d0d64c51b79f5baf",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post Review data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostReviewResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (Validation error)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
        "/api/v1/post-review/{id}": {
            "put": {
                "tags": [
                    "Post Review"
                ],
                "summary": "Update an existing post review",
                "description": "Update the specified resource in storage.",
                "operationId": "87fc339da83c7e2d14925773ec1d8ebe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post review to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post Review data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated post review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostReviewResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (Validation error)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post Review"
                ],
                "summary": "Delete an post review",
                "description": "Remove the specified resource from storage.",
                "operationId": "c23a2086c2d116f72c91fb0038fc3967",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post review to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted post review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post Review deleted successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PostReviewResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "An object representing a validation error response",
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "An object containing field validation error messages. (This may varies from your api endpoints.)",
                        "type": "object",
                        "example": {
                            "postkey": [
                                "The postkey field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "StorePostReviewRequest": {
                "title": "Store Post Review Request",
                "description": "Request data for storing post reviews",
                "required": [
                    "post_id",
                    "reviews"
                ],
                "properties": {
                    "post_id": {
                        "description": "The ID of the post",
                        "type": "integer"
                    },
                    "reviews": {
                        "description": "Array of post review details",
                        "type": "array",
                        "items": {
                            "title": "Post Review",
                            "description": "Post review details",
                            "required": [
                                "title",
                                "amount"
                            ],
                            "properties": {
                                "review_id": {
                                    "description": "The id of the review",
                                    "type": "integer",
                                    "example": "1"
                                },
                                "review": {
                                    "description": "The review of the post",
                                    "type": "string",
                                    "example": "Impressive."
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdatePostReviewRequest": {
                "title": "Post Review  Update Request",
                "description": "Post Review  update request data",
                "required": [
                    "review"
                ],
                "properties": {
                    "review": {
                        "description": "The Review of the  Post.",
                        "type": "string",
                        "example": "Very Impressive"
                    }
                },
                "type": "object"
            },
...
            "PostReviewResource": {
                "title": "Post Review Resource",
                "description": "Post Review resource representation",
                "properties": {
                    "id": {
                        "description": "The unique id of the post review",
                        "type": "integer",
                        "example": "1"
                    },
                    "post_id": {
                        "description": "The Id of the post ",
                        "type": "integer",
                        "example": "1"
                    },
                    "review_id": {
                        "description": "The Id of the Review",
                        "type": "integer",
                        "example": "1"
                    },
                    "review": {
                        "description": "The review of the post",
                        "type": "string",
                        "example": "Very Nice."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Bearer Token Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}